buyfast/requirements.txt
flask
pymongo
ollama
Pillow
crewai
selenium
